+++ packages/domain-events-infra/linter.tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../node_modules/@standardnotes/config/src/linter.tsconfig.json"
+}
