+++ packages/domain-events-infra/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@standardnotes/domain-events",
-  "version": "2.0.0",
+  "name": "@standardnotes/domain-events-infra",
+  "version": "1.0.0",
   "engines": {
     "node": ">=14.0.0 <16.0.0"
   },
-  "description": "Domain Events SDK used in SN projects",
+  "description": "Domain Events Infrastructure SDK used in SN projects",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
@@ -20,13 +20,24 @@
     "prebuild": "yarn clean",
     "build": "tsc -p tsconfig.json",
     "lint": "eslint . --ext .ts",
-    "test:unit": "jest spec --coverage --passWithNoTests"
+    "test:unit": "jest spec --coverage"
   },
   "dependencies": {
-    "@standardnotes/auth": "^3.2.0"
+    "@standardnotes/domain-events": "2.0.0",
+    "aws-sdk": "^2.824.0",
+    "ioredis": "^4.19.4",
+    "newrelic": "^7.4.0",
+    "reflect-metadata": "^0.1.13",
+    "sqs-consumer": "^5.5.0",
+    "winston": "^3.3.3"
   },
   "devDependencies": {
     "@standardnotes/config": "^1.2.0",
+    "@types/ioredis": "^4.17.10",
+    "@types/jest": "^26.0.15",
+    "@types/newrelic": "^7.0.1",
+    "jest": "^26.6.3",
+    "ts-jest": "^26.4.4",
     "typescript": "^4.0.5"
   },
   "jest": {
