+++ packages/domain-events-infra/src/index.ts
@@ -1,0 +1,1 @@
+export * from './Infra'
