+++ packages/domain-events-infra/src/test-setup.ts
@@ -1,0 +1,2 @@
+process.env.NEW_RELIC_ENABLED = 'false'
+process.env.NEW_RELIC_NO_CONFIG_FILE = 'true'
