+++ packages/domain-events/src/index.ts
@@ -1,2 +1,1 @@
 export * from './Domain'
-export * from './Infra'
