--- packages/domain-events/src/test-setup.ts
@@ -1,2 +1,0 @@
-process.env.NEW_RELIC_ENABLED = 'false'
-process.env.NEW_RELIC_NO_CONFIG_FILE = 'true'
