+++ packages/snjs/package.json
@@ -75,7 +75,7 @@
   "dependencies": {
     "@standardnotes/auth": "3.1.1",
     "@standardnotes/common": "1.1.0",
-    "@standardnotes/domain-events": "1.9.0",
+    "@standardnotes/domain-events": "2.0.0",
     "@standardnotes/features": "1.4.0",
     "@standardnotes/settings": "1.2.0",
     "@standardnotes/sncrypto-common": "1.5.2"
